corrected some logic mistakes with err_code
