+++ task_02/parallel_work.cpp
@@ -8,27 +8,6 @@
 #define unlikely(expr) __builtin_expect(!!(expr), 0)
 #define likely(expr) __builtin_expect(!!(expr), 1)
 
-void err_message(int code) {
-  switch (code) {
-    case NO_FILENAME:
-      printf("No filename!\n");
-      break;
-    case ERR_FILE:
-      printf("Couldn't open file!\n");
-      break;
-    case MEM_ERR:
-      printf("Memory allocation error!\n");
-      break;
-    case ZERO_MAINSTR:
-      printf("String where we find substring has zero length!\n");
-      break;
-    case ZERO_SUBSTR:
-      printf("Substring has zero length!\n");
-  default:
-      printf("Error!\n");
-  } 
-}
-
 void *thread_func(void *args) {
   task_args *arg = (task_args *) args;
   int len_str = strlen(arg->str);
